added signup functionality
edited the port in SignUp.js
added function signupHandler to api.go

FE asks for first name and last name. combined them with a space separation and thats what gets appended to DB.

user.role is "None" by default/hardcoded. will handle the change of the role in another ticket for when the admin goes to manage users.