In this commit there was a lot added, the HVAC page now has a thermometer that shows the temp in the room and it can be adjusted, along with different modes and fan speed.
The Energy page is now able to update the status of all the devices and it will be sent to iot.

The Security page, now has a keypad that we can use to arm and disarm a security system, the functionality is still in progress. Also a lock animation was added to show when it was locked vs unlocked.

 When the website gets scaled a hamburger menu would popup but had no functionality in navigating the user, this is now fixed and and it gives the user a menu to navigate through the pages.